Distance Sensor constant control; cacheGoals
distSensorControl function in movement file controls whether distance sensor is UP or DOWN, assumes UP at beginning of running the program.
cacheGoals is supposed to take the goals that the opponent has scored and send to other bot. Need to test whether function to determine scored works, need to see if array converts to string properly.